+++ assets/browserify/admin/form-editor.js
@@ -100,11 +100,13 @@ function updatePreview() {
 
     // replace template tags
     markup = markup.replace(templateRegex, function(s, m) {
+
+        // if a default value was provided, use that
         if(arguments[3]) {
             return arguments[3];
         }
 
-        return '';
+        return '';        
     });
 
     // update dom
@@ -157,5 +159,6 @@ function uniq(a) {
 
 export default {
     init,
-    replaceSelection
+    replaceSelection,
+    updateFieldVariables,
 };