+++ src/Actions/Email.php
@@ -70,7 +70,7 @@ class Email extends Action {
                <th><label><?php echo __( 'Content Type', 'html-forms' ); ?></label></th>
                <td>
                    <select name="form[settings][actions][<?php echo $index; ?>][content_type]" required>
-                      <option <?php selected( $settings['content_type'], 'text/html' ); ?>>text/plain</option>
+                      <option <?php selected( $settings['content_type'], 'text/plain' ); ?>>text/plain</option>
                       <option <?php selected( $settings['content_type'], 'text/html' ); ?>>text/html</option>
                    </select>
                </td>
@@ -100,24 +100,23 @@ class Email extends Action {
 
     $settings = array_merge( $this->get_default_settings(), $settings );
 
-    $to = hf_replace_data_variables( $settings['to'], $submission->data );
-    $subject = ! empty( $settings['subject'] ) ? hf_replace_data_variables( $settings['subject'], $submission->data ) : '';
-    $message = hf_replace_data_variables( $settings['message'], $submission->data );
+    $to = hf_replace_data_variables( $settings['to'], $submission->data, 'strip_tags' );
+    $subject = ! empty( $settings['subject'] ) ? hf_replace_data_variables( $settings['subject'], $submission->data, 'strip_tags' ) : '';
+    $message = hf_replace_data_variables( $settings['message'], $submission->data, 'strip_tags' );
     
     // parse additional email headers from settings
     $headers = array();
     if( ! empty( $settings['headers'] ) ) {
-      $headers = explode( PHP_EOL, hf_replace_data_variables( $settings['headers'], $submission->data ) );
+      $headers = explode( PHP_EOL, hf_replace_data_variables( $settings['headers'], $submission->data, 'strip_tags' ) );
     }
 
     $html_email = $settings['content_type'] === 'text/html';
     if( $html_email ) {
       $headers[] = 'Content-Type: text/html';
-      $message = nl2br( $message );
     } 
 
     if( ! empty( $settings['from'] ) ) {
-      $from = hf_replace_data_variables($settings['from'], $submission->data);
+      $from = hf_replace_data_variables($settings['from'], $submission->data, 'strip_tags');
       $headers[] = sprintf( 'From: %s', $from );
     }